pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Complete ML Batch Job, including MLOps Best Practices"
authors = ["Ezzaldin97 <kingtal632@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
python-dotenv = "^1.0.0"
pandas = "^2.0.1"
requests = "^2.30.0"
duckdb = "^0.9.2"
ploomber-cloud = "^0.1.1"
scikit-learn = "^1.3.0"
lightgbm = "^3.3.0"
prefect = "^2.10.0"
streamlit = "^1.20.0"
hyperopt = "^0.2.7"
prefect-email = "^0.3.2"
sktime = "^0.24.1"
pydantic = "^2.5.0"
geopy = "^2.4.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.1"


[tool.poetry.group.linting.dependencies]
black = "^23.3.0"
ruff = "^0.0.287"

[tool.black]
line-length = 88
target_version = ['py310', 'py311', 'py312']
preview = true

[tool.ruff]
line-length = 88
src = ["app"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
